refactor:코드 컴포넌트별 분리 및 작성성form에 사진 띄우기
