+++ src/components/Community/Header2.jsx
@@ -4,8 +4,9 @@ import { IoIosArrowBack } from "react-icons/io";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import { RowContainer } from "../../styles/GlobalStyle";
+import photoIcon from "../../images/photoIcon.png";
 
-export default function Header2({ isDrop }) {
+export default function Header2({ isDrop, handleUpload, isPictureIcon }) {
   const navigate = useNavigate();
   const handleBack = () => {
     navigate(-1);
@@ -20,6 +21,22 @@ export default function Header2({ isDrop }) {
       <RowContainer style={{ padding: `${responsiveSize("20")}` }}>
         <ResponsiveIcon onClick={handleBack} />
         {isDrop ? <DropDown /> : undefined}
+        {/* 사진 올리는 상단바 */}
+        {isPictureIcon ? (
+          <RowContainer style={{ padding: `${responsiveSize("20")}` }}>
+            <input
+              type="file"
+              name="photo"
+              accept="image/*,audio/*,video/mp4,video/x-m4v"
+              onChange={handleUpload}
+              style={{ display: "none" }}
+              id="file-upload"
+            />
+            <label htmlFor="file-upload" style={{ cursor: "pointer" }}>
+              <img src={photoIcon} alt="photoicon" />
+            </label>
+          </RowContainer>
+        ) : undefined}
       </RowContainer>
     </>
   );
